Move function to context
