+++ feedback-app/src/App.js
@@ -1,4 +1,3 @@
-import { useState } from "react";
 import {BrowserRouter as Router, Route, Routes} from 'react-router-dom'
 
 import { FeedbackProvider } from "./context/FeedbackContext";
@@ -7,25 +6,11 @@ import Header from "./components/Header";
 import FeedbackStats from "./components/FeedbackStats";
 import FeedbackList from "./components/FeedbackList";
 import FeedbackForm from "./components/FeedbackForm";
-import FeedbackData from "./data/FeedbackData";
 
 import AboutPage from "./pages/AboutPage";
 import AboutIconLink from "./components/AboutIconLink";
 
 function App(){
-    const [feedback, setFeedback] = useState( FeedbackData)
-
-    const addFeedback = (newFeedback) => {
-        let maxId = Math.max(...feedback.map( (item) => (item.id)))
-        newFeedback.id = maxId + 1
-        setFeedback([newFeedback, ...feedback])
-    }
-    
-    const handleDelete = (id) => {
-        if (window.confirm(`Are you sure you want to delete ${id} item?`)){
-            setFeedback(feedback.filter((item) => item.id !== id))
-        }
-    }
 
     return (
         <FeedbackProvider>
@@ -38,9 +23,9 @@ function App(){
                     path='/'
                     element={
                         <>
-                            <FeedbackForm handleAdd={addFeedback} />
+                            <FeedbackForm />
                             <FeedbackStats />
-                            <FeedbackList handleDelete={handleDelete}/>
+                            <FeedbackList />
                         </>
                     }>  
                     </Route>
