+++ feedback-app/src/components/FeedbackItem.jsx
@@ -1,9 +1,11 @@
 import {FaTimes} from 'react-icons/fa'
+import { useContext } from 'react'
+import FeedbackContext from '../context/FeedbackContext'
 import PropTypes from 'prop-types'
 import Card from "./shared/Card"
 
-function FeedbackItem({item, handleDelete}){
-
+function FeedbackItem({item }){
+    const {handleDelete} = useContext(FeedbackContext)
 
     return (
         <Card>
