+++ feedback-app/src/data/FeedbackData.jsx
@@ -9,11 +9,6 @@ const FeedbackData = [
       rating: 9,
       text: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. consequuntur vel vitae commodi alias voluptatem est voluptatum ipsa quae.',
     },
-    {
-      id: 3,
-      rating: 8,
-      text: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. consequuntur vel vitae commodi alias voluptatem est voluptatum ipsa quae.',
-    },
   ]
   
   export default FeedbackData
